wallet help text/error messages updated
Corrected error messages with a terminating period for log reading clarity. Capitalized beginning of sentences.
Provided additional descriptions for help commands for start_mining, transfer, locked_transfer, viewkey, spendkey.
Added description of when creating a new wallet that there is a wallet file and a .key files so the user knows to keep both files.
Explained that there is no recovery option for losing your mnemonic seed or keys when the wallet is created.